chore(auth): flatten auth flow structure
Generated-By: PostHog Desktop
Task-Id: cded1552-8ba7-4e00-aa7d-a2e57f7a453a
