+++ frontend/src/scenes/authentication/login/Login.stories.tsx
@@ -1,165 +1,133 @@
-import { Meta, StoryObj } from '@storybook/react'
-import { router } from 'kea-router'
+import type { Meta, StoryFn } from '@storybook/react'
+import { useEffect } from 'react'
 
-import { useDelayedOnMountEffect } from 'lib/hooks/useOnMountEffect'
-import { Login2FA } from 'scenes/authentication/login-2fa/Login2FA'
-import { urls } from 'scenes/urls'
+import { useStorybookMocks } from '~/mocks/browser'
+import preflightJson from '~/mocks/fixtures/_preflight.json'
 
-import { mswDecorator, useStorybookMocks } from '~/mocks/browser'
-
-import preflightJson from '../../../mocks/fixtures/_preflight.json'
 import { Login } from './Login'
 import { loginLogic } from './loginLogic'
 
-const meta: Meta = {
+type StoryArgs = {
+    cloud: boolean
+    region: 'US' | 'EU'
+    googleOAuth: boolean
+    github: boolean
+    gitlab: boolean
+    samlAvailable: boolean
+    ssoEnforcement: 'none' | 'google-oauth2' | 'github' | 'gitlab' | 'saml'
+    generalError: 'none' | 'invalid_credentials' | 'code_based_verification_sent'
+}
+
+const meta: Meta<StoryArgs> = {
     title: 'Scenes-Other/Login',
+    tags: ['test-skip'],
     parameters: {
         layout: 'fullscreen',
         viewMode: 'story',
     },
-    decorators: [
-        mswDecorator({
-            post: {
-                '/api/login/precheck': { sso_enforcement: null, saml_available: false },
-            },
-        }),
-    ],
+    argTypes: {
+        cloud: { control: 'boolean', name: 'Cloud' },
+        region: { control: 'select', options: ['US', 'EU'], name: 'Region', if: { arg: 'cloud' } },
+        googleOAuth: { control: 'boolean', name: 'Google OAuth' },
+        github: { control: 'boolean', name: 'GitHub' },
+        gitlab: { control: 'boolean', name: 'GitLab' },
+        samlAvailable: { control: 'boolean', name: 'SAML available' },
+        ssoEnforcement: {
+            control: 'select',
+            name: 'SSO enforcement',
+            options: ['none', 'google-oauth2', 'github', 'gitlab', 'saml'],
+        },
+        generalError: {
+            control: 'select',
+            name: 'General error',
+            options: ['none', 'invalid_credentials', 'code_based_verification_sent'],
+        },
+    },
+    args: {
+        cloud: true,
+        region: 'US',
+        googleOAuth: true,
+        github: true,
+        gitlab: true,
+        samlAvailable: false,
+        ssoEnforcement: 'none',
+        generalError: 'none',
+    },
 }
 export default meta
 
-type Story = StoryObj<{}>
+const Template: StoryFn<StoryArgs> = ({
+    cloud,
+    region,
+    googleOAuth,
+    github,
+    gitlab,
+    samlAvailable,
+    ssoEnforcement,
+    generalError,
+}) => {
+    const enforcement = ssoEnforcement === 'none' ? null : ssoEnforcement
 
-export const Cloud: Story = {
-    render: () => {
-        useStorybookMocks({
-            get: {
-                '/_preflight': {
-                    ...preflightJson,
-                    cloud: true,
-                    realm: 'cloud',
-                    can_create_org: true,
-                    available_social_auth_providers: { github: true, gitlab: true, 'google-oauth2': true, saml: false },
+    useStorybookMocks({
+        get: {
+            '/_preflight': {
+                ...preflightJson,
+                cloud,
+                region: cloud ? region : undefined,
+                realm: cloud ? 'cloud' : 'hosted-clickhouse',
+                is_debug: cloud,
+                can_create_org: cloud,
+                available_social_auth_providers: {
+                    'google-oauth2': googleOAuth,
+                    github,
+                    gitlab,
+                    saml: samlAvailable,
                 },
             },
-        })
-        return <Login />
-    },
-}
+        },
+        post: {
+            '/api/login/precheck': { sso_enforcement: enforcement, saml_available: samlAvailable },
+        },
+    })
 
-export const CloudEU: Story = {
-    render: () => {
-        useStorybookMocks({
-            get: {
-                '/_preflight': {
-                    ...preflightJson,
-                    cloud: true,
-                    region: 'EU',
-                    realm: 'cloud',
-                    can_create_org: true,
-                    available_social_auth_providers: { github: true, gitlab: true, 'google-oauth2': true, saml: false },
-                },
-            },
-        })
-        return <Login />
-    },
-}
-
-export const CloudWithGoogleLoginEnforcement: Story = {
-    render: () => {
-        useStorybookMocks({
-            get: {
-                '/_preflight': {
-                    ...preflightJson,
-                    cloud: true,
-                    realm: 'cloud',
-                    can_create_org: true,
-                    available_social_auth_providers: { github: true, gitlab: true, 'google-oauth2': true, saml: false },
-                },
-            },
-            post: {
-                '/api/login/precheck': { sso_enforcement: 'google-oauth2', saml_available: false },
-            },
-        })
-
-        // Trigger pre-check
-        useDelayedOnMountEffect(() => {
+    useEffect(() => {
+        if (enforcement) {
             loginLogic.actions.setLoginValue('email', 'test@posthog.com')
             loginLogic.actions.precheck({ email: 'test@posthog.com' })
-        })
+        }
+    }, [enforcement])
 
-        return <Login />
-    },
-    parameters: {
-        testOptions: {
-            waitForSelector: '[href^="/login/google-oauth2/"]',
-        },
-    },
+    useEffect(() => {
+        if (generalError !== 'none') {
+            const messages: Record<string, string> = {
+                invalid_credentials: 'Invalid email or password.',
+                code_based_verification_sent: 'Check your email to verify your account.',
+            }
+            loginLogic.actions.setGeneralError(generalError, messages[generalError] ?? '')
+        } else {
+            loginLogic.actions.clearGeneralError()
+        }
+    }, [generalError])
+
+    return <Login />
 }
 
-export const SelfHosted: Story = {
-    render: () => {
-        useStorybookMocks({
-            get: {
-                '/_preflight': {
-                    ...preflightJson,
-                    cloud: false,
-                    realm: 'hosted-clickhouse',
-                    available_social_auth_providers: {
-                        github: false,
-                        gitlab: false,
-                        'google-oauth2': false,
-                        saml: false,
-                    },
-                },
-            },
-        })
-        return <Login />
-    },
-}
+export const Default: StoryFn<StoryArgs> = Template.bind({})
 
-export const SelfHostedWithSAML: Story = {
-    render: () => {
-        useStorybookMocks({
-            get: {
-                '/_preflight': {
-                    ...preflightJson,
-                    cloud: false,
-                    realm: 'hosted-clickhouse',
-                    available_social_auth_providers: {
-                        github: false,
-                        gitlab: false,
-                        'google-oauth2': false,
-                        saml: true,
-                    },
-                },
-            },
-        })
-        return <Login />
-    },
-    parameters: {
-        testOptions: {
-            waitForSelector: '[href^="/login/saml/"]',
-        },
-    },
-}
+export const SelfHosted: StoryFn<StoryArgs> = Template.bind({})
+SelfHosted.args = { cloud: false, googleOAuth: false, github: false, gitlab: false, samlAvailable: false }
 
-export const SSOError: Story = {
-    render: () => {
-        useStorybookMocks({
-            get: {
-                '/_preflight': preflightJson,
-            },
-        })
+export const CloudEU: StoryFn<StoryArgs> = Template.bind({})
+CloudEU.args = { region: 'EU' }
 
-        useDelayedOnMountEffect(() => router.actions.push(`${urls.login()}?error_code=improperly_configured_sso`))
+export const SSOEnforced: StoryFn<StoryArgs> = Template.bind({})
+SSOEnforced.args = { ssoEnforcement: 'google-oauth2' }
 
-        return <Login />
-    },
-}
+export const SAMLAvailable: StoryFn<StoryArgs> = Template.bind({})
+SAMLAvailable.args = { samlAvailable: true }
 
-export const SecondFactor: Story = {
-    render: () => {
-        useDelayedOnMountEffect(() => router.actions.push(urls.login2FA()))
-        return <Login2FA />
-    },
-}
+export const LoginError: StoryFn<StoryArgs> = Template.bind({})
+LoginError.args = { generalError: 'invalid_credentials' }
+
+export const EmailVerification: StoryFn<StoryArgs> = Template.bind({})
+EmailVerification.args = { generalError: 'code_based_verification_sent' }
