+++ frontend/src/scenes/authentication/shared/authScene/AuthScene.scss
@@ -1,9 +1,9 @@
-// Paper-desk auth variant — light mode only (unauthenticated auth scenes are forced light).
+// Unauthenticated auth scenes are forced light.
 // Only rules that cannot be expressed in Tailwind utilities live here:
 // pseudo-elements, interactive pseudo-class combinators, child selectors, keyframes.
 
 // Dotted paper grid + faint accent glow — multi-layer background (per-layer size/repeat), not achievable via Tailwind
-.PaperDesk {
+.AuthScene {
     --primary-3000-button-bg: var(--warning);
     --primary-3000-frame-bg: var(--primary-3000-button-border-light);
     --primary-3000-button-border: var(--primary-3000-button-border-light);
@@ -14,7 +14,7 @@
 
     // Accent glow over the dotted grid. Layered onto the element itself rather than absolutely
     // positioned pseudo-elements so the texture stays fixed to the viewport while the scene scrolls
-    // (a card taller than the viewport scrolls inside .PaperDesk; pseudo-elements would scroll away).
+    // (a card taller than the viewport scrolls inside .AuthScene; pseudo-elements would scroll away).
     // background-color stays on the Tailwind `bg-[#eef0e7]` utility — only the image layers live here.
     background-image:
         radial-gradient(72% 72% at 50% 38%, rgb(247 165 1 / 5%), transparent 70%),
@@ -25,10 +25,10 @@
         16px 16px;
 }
 
-// Glass variant uses posthog.com's layered garden greens with a soft light well behind the card.
+// The auth scene uses layered garden greens with a soft light well behind the card.
 // The base color overrides the parchment `bg-[#eef0e7]` utility because unlayered rules beat
 // Tailwind's cascade layer.
-.PaperDesk--glass {
+.AuthScene--garden {
     --color-text-warning: var(--color-text-primary);
 
     background-color: #b5cd8f;
@@ -39,7 +39,7 @@
     background-size: auto;
 
     // The cream OS-window card the new website (and the self-driving onboarding) uses.
-    .PaperDesk__card {
+    .AuthScene__card {
         background: #f6f5f0;
         border-color: #e3e0d4;
         border-radius: 1rem;
@@ -47,13 +47,13 @@
     }
 
     // Corner notes read as dark-green ink on the garden rather than gray on parchment.
-    .PaperDesk__notes div {
+    .AuthScene__notes div {
         color: rgb(45 74 22 / 65%);
     }
 }
 
 // Paper sheet resting on the desk — the login card and the dev login panel share it.
-.PaperDesk__card {
+.AuthScene__card {
     background: #fff;
     border: 1px solid #e0e1d9;
     border-radius: var(--radius-lg);
@@ -63,7 +63,7 @@
 }
 
 // Child SVG sizing — child selector, no Tailwind equivalent
-.PaperDesk__logo {
+.AuthScene__logo {
     svg {
         display: block;
         width: auto;
@@ -76,7 +76,7 @@
 }
 
 // Invite header mark SVG sizing
-.PaperDesk__inviteHeader-mark {
+.AuthScene__inviteHeader-mark {
     svg {
         width: auto;
         height: 30px;
@@ -84,9 +84,9 @@
 }
 
 // Sub paragraph — code and mono spans inside it share styles
-.PaperDesk__sub {
+.AuthScene__sub {
     code,
-    .PaperDesk__mono {
+    .AuthScene__mono {
         font-family: var(--font-mono);
         font-size: 13px;
         color: #111;
@@ -94,7 +94,7 @@
 }
 
 // Terms text — nested link style
-.PaperDesk__terms {
+.AuthScene__terms {
     a {
         font-weight: 600;
         color: rgb(17 17 17 / 58%);
@@ -104,34 +104,34 @@
 }
 
 // Note callout — bold inside it is darker
-.PaperDesk__note {
+.AuthScene__note {
     b {
         color: #111;
     }
 }
 
 // Blink animation for typewriter caret
-.PaperDesk__caret {
-    animation: PaperDesk__blink 1.05s steps(1) infinite;
+.AuthScene__caret {
+    animation: AuthScene__blink 1.05s steps(1) infinite;
 }
 
 @media (prefers-reduced-motion: no-preference) {
-    .PaperDesk__column {
-        animation: PaperDesk__rise 0.28s cubic-bezier(0.16, 1, 0.3, 1);
+    .AuthScene__column {
+        animation: AuthScene__rise 0.28s cubic-bezier(0.16, 1, 0.3, 1);
     }
 
-    .PaperDesk__progress-fill {
-        animation: PaperDesk__progress 2.2s cubic-bezier(0.16, 1, 0.3, 1);
+    .AuthScene__progress-fill {
+        animation: AuthScene__progress 2.2s cubic-bezier(0.16, 1, 0.3, 1);
     }
 }
 
 @media (prefers-reduced-motion: reduce) {
-    .PaperDesk__caret {
+    .AuthScene__caret {
         animation: none;
     }
 }
 
-@keyframes PaperDesk__blink {
+@keyframes AuthScene__blink {
     0%,
     50% {
         opacity: 1;
@@ -143,7 +143,7 @@
     }
 }
 
-@keyframes PaperDesk__rise {
+@keyframes AuthScene__rise {
     from {
         opacity: 0;
         transform: translateY(8px);
@@ -155,7 +155,7 @@
     }
 }
 
-@keyframes PaperDesk__progress {
+@keyframes AuthScene__progress {
     from {
         width: 0%;
     }
