+++ frontend/src/scenes/authentication/shared/authScene/Typewriter.tsx
@@ -37,7 +37,7 @@ export function Typewriter({ lines }: { lines: string[] }): JSX.Element {
     const parts = full.slice(0, shown).split('\n')
     return (
         <div
-            className="PaperDesk__notes absolute top-[clamp(20px,4vh,40px)] left-[clamp(20px,4vw,44px)] z-[2] min-h-[2.6em] font-mono text-xs leading-relaxed whitespace-pre"
+            className="AuthScene__notes absolute top-[clamp(20px,4vh,40px)] left-[clamp(20px,4vw,44px)] z-[2] min-h-[2.6em] font-mono text-xs leading-relaxed whitespace-pre"
             aria-hidden
             // Chrome's in-page translation replaces text nodes with <font> elements; on this
             // 40ms-tick animation React then crashes removing text it no longer owns
@@ -48,7 +48,7 @@ export function Typewriter({ lines }: { lines: string[] }): JSX.Element {
                 <div key={idx} className="text-primary/50 first:font-semibold">
                     {part}
                     {idx === parts.length - 1 && (
-                        <span className="PaperDesk__caret inline-block w-2 h-[1.05em] ml-0.5 [vertical-align:-2px] bg-warning rounded-[1px]" />
+                        <span className="AuthScene__caret inline-block w-2 h-[1.05em] ml-0.5 [vertical-align:-2px] bg-warning rounded-[1px]" />
                     )}
                 </div>
             ))}
