--- frontend/src/scenes/authentication/signup/SignupContainer.tsx
@@ -1,11 +1,0 @@
-import { SceneExport } from 'scenes/sceneTypes'
-
-import { GlassSignup } from './variants/glass/GlassSignup'
-
-export const scene: SceneExport = {
-    component: SignupContainer,
-}
-
-export function SignupContainer(): JSX.Element | null {
-    return <GlassSignup />
-}
