+++ frontend/src/scenes/authentication/verify-email/VerifyEmail.stories.tsx
@@ -1,63 +1,65 @@
-import type { Meta, StoryObj } from '@storybook/react'
+import type { Meta, StoryFn } from '@storybook/react'
+import { useEffect } from 'react'
 
-import { useDelayedOnMountEffect } from 'lib/hooks/useOnMountEffect'
+import { useStorybookMocks } from '~/mocks/browser'
+import preflightJson from '~/mocks/fixtures/_preflight.json'
 
 import { VerifyEmail } from './VerifyEmail'
 import { verifyEmailLogic } from './verifyEmailLogic'
 
-const meta: Meta = {
+type VerifyEmailView = 'pending' | 'verify' | 'success' | 'invalid'
+
+type StoryArgs = {
+    view: VerifyEmailView
+}
+
+const meta: Meta<StoryArgs> = {
     title: 'Scenes-Other/Verify Email',
+    tags: ['test-skip'],
     parameters: {
         layout: 'fullscreen',
         viewMode: 'story',
     },
-}
-export default meta
-type Story = StoryObj<{}>
-
-export const VerifyEmailPending: Story = {
-    render: () => {
-        useDelayedOnMountEffect(() => {
-            verifyEmailLogic.actions.setView('pending')
-            verifyEmailLogic.actions.setUuid('12345678')
-        })
-
-        return <VerifyEmail />
-    },
-}
-
-export const VerifyingEmail: Story = {
-    render: () => {
-        useDelayedOnMountEffect(() => {
-            verifyEmailLogic.actions.setView('verify')
-        })
-
-        return <VerifyEmail />
-    },
-    parameters: {
-        testOptions: {
-            waitForLoadersToDisappear: false,
+    argTypes: {
+        view: {
+            control: 'select',
+            name: 'View',
+            options: ['pending', 'verify', 'success', 'invalid'] satisfies VerifyEmailView[],
         },
     },
-}
-
-export const VerifyEmailSuccess: Story = {
-    render: () => {
-        useDelayedOnMountEffect(() => {
-            verifyEmailLogic.actions.setView('success')
-        })
-
-        return <VerifyEmail />
+    args: {
+        view: 'pending',
     },
 }
+export default meta
 
-export const VerifyEmailInvalid: Story = {
-    render: () => {
-        useDelayedOnMountEffect(() => {
-            verifyEmailLogic.actions.setView('invalid')
-            verifyEmailLogic.actions.setUuid('12345678')
-        })
+const Template: StoryFn<StoryArgs> = ({ view }) => {
+    useStorybookMocks({
+        get: {
+            '/_preflight': { ...preflightJson },
+            '/api/users/@me': () => [200, { email: 'test@posthog.com', first_name: 'Test' }],
+        },
+    })
 
-        return <VerifyEmail />
-    },
+    useEffect(() => {
+        verifyEmailLogic.actions.setView(view)
+        verifyEmailLogic.actions.setUuid('12345678')
+    }, [view])
+
+    return <VerifyEmail />
 }
+
+export const Default: StoryFn<StoryArgs> = Template.bind({})
+
+export const Pending: StoryFn<StoryArgs> = Template.bind({})
+Pending.args = { view: 'pending' }
+
+export const Verifying: StoryFn<StoryArgs> = Template.bind({})
+Verifying.args = { view: 'verify' }
+Verifying.parameters = { testOptions: { waitForLoadersToDisappear: false } }
+
+export const Success: StoryFn<StoryArgs> = Template.bind({})
+Success.args = { view: 'success' }
+
+export const InvalidLink: StoryFn<StoryArgs> = Template.bind({})
+InvalidLink.args = { view: 'invalid' }
