+++ frontend/src/scenes/onboarding/self-driving/SelfDrivingOnboarding.scss
@@ -3,10 +3,7 @@
     background-color: var(--color-bg-surface-primary);
 }
 
-// `sm`+ (≥640px): dotted paper grid + faint amber glow, copied from the glass auth experience
-// (scenes/authentication/shared/paperDesk/PaperDesk.scss) so onboarding and sign-in/sign-up share the
 // exact same backdrop. Multi-layer background (per-layer size/repeat), not achievable via Tailwind.
-// Unlike the glass auth experience (light-only), the base color lives here too so it can flip with the theme without
 // competing against a Tailwind `bg-*` utility's cascade layer.
 @media (min-width: 640px) {
     // Experiment: posthog.com's garden look - layered greens with a soft light well behind the
@@ -24,7 +21,6 @@
             16px 16px;
     }
 
-    // Onboarding also runs in dark mode (the glass auth experience is light-only), so swap the parchment base and
     // invert the dot tint there. The grid geometry (1.4px dots, 16px cell) and amber glow are unchanged.
     [theme='dark'] .OnboardingDottedBg {
         background-color: var(--color-bg-surface-tertiary);
